Added Database integration for index page. Made all include paths absolute.
index.php now checks get parameters skip and board.
If the board is set, the page will display posts from that board only.
if the board is not set, the page will just display the most recent posts from all boards.
The skip parameter skips the specified number of posts when selecting which posts should be displayed. If there are enough posts to warrant a second page, we would implement that with the skip parameter (essentially 'skipping' the posts of the first page).